Updated dependency for aws-cloudformation-rpdk-java-plugin
Updated dependency for aws-cloudformation-rpdk-java-plugin
