+++ namedquery/pom.xml
@@ -32,9 +32,9 @@
             <artifactId>athena</artifactId>
             <version>2.8.7</version>
         </dependency>
-        <!-- https://github.com/awslabs/aws-cloudformation-rpdk-java-plugin/ -->
+        <!-- https://github.com/aws-cloudformation/aws-cloudformation-rpdk-java-plugin/ -->
         <dependency>
-            <groupId>com.amazonaws.cloudformation</groupId>
+            <groupId>software.amazon.cloudformation</groupId>
             <artifactId>aws-cloudformation-rpdk-java-plugin</artifactId>
             <version>1.0</version>
         </dependency>
