--- namedquery/src/main/java/com/amazonaws/athena/namedquery/CallbackContext.java
@@ -1,4 +1,0 @@
-package com.amazonaws.athena.namedquery;
-
-public class CallbackContext {
-}
