--- namedquery/src/main/java/com/amazonaws/athena/namedquery/Configuration.java
@@ -1,26 +1,0 @@
-package com.amazonaws.athena.namedquery;
-
-import java.util.Map;
-
-import org.json.JSONObject;
-import org.json.JSONTokener;
-
-class Configuration extends BaseConfiguration {
-
-    public Configuration() {
-        super("aws-athena-namedquery.json");
-    }
-
-    public JSONObject resourceSchemaJSONObject() {
-        return new JSONObject(new JSONTokener(this.getClass().getClassLoader().getResourceAsStream(schemaFilename)));
-    }
-
-    /**
-     * Providers should implement this method if their resource has a 'Tags' property to define resource-level tags
-     *
-     * @return
-     */
-    public Map<String, String> resourceDefinedTags(final ResourceModel resourceModel) {
-        return null;
-    }
-}
