+++ namedquery/src/main/java/software/amazon/athena/namedquery/CallbackContext.java
@@ -1,0 +1,4 @@
+package software.amazon.athena.namedquery;
+
+public class CallbackContext {
+}
