+++ namedquery/src/main/java/software/amazon/athena/namedquery/Configuration.java
@@ -1,0 +1,8 @@
+package software.amazon.athena.namedquery;
+
+class Configuration extends BaseConfiguration {
+
+    public Configuration() {
+        super("aws-athena-namedquery.json");
+    }
+}
