+++ namedquery/src/main/java/software/amazon/athena/namedquery/CreateHandler.java
@@ -1,17 +1,15 @@
-package com.amazonaws.athena.namedquery;
-
-import com.amazonaws.cloudformation.exceptions.CfnGeneralServiceException;
-import com.amazonaws.cloudformation.exceptions.CfnInvalidRequestException;
-import com.amazonaws.cloudformation.proxy.AmazonWebServicesClientProxy;
-import com.amazonaws.cloudformation.proxy.Logger;
-import com.amazonaws.cloudformation.proxy.OperationStatus;
-import com.amazonaws.cloudformation.proxy.ProgressEvent;
-import com.amazonaws.cloudformation.proxy.ResourceHandlerRequest;
+package software.amazon.athena.namedquery;
 
 import software.amazon.awssdk.services.athena.AthenaClient;
 import software.amazon.awssdk.services.athena.model.CreateNamedQueryRequest;
 import software.amazon.awssdk.services.athena.model.InternalServerException;
 import software.amazon.awssdk.services.athena.model.InvalidRequestException;
+import software.amazon.cloudformation.exceptions.CfnGeneralServiceException;
+import software.amazon.cloudformation.exceptions.CfnInvalidRequestException;
+import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
+import software.amazon.cloudformation.proxy.Logger;
+import software.amazon.cloudformation.proxy.ProgressEvent;
+import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
 
 public class CreateHandler extends BaseHandler<CallbackContext> {
     private AmazonWebServicesClientProxy clientProxy;
@@ -32,10 +30,7 @@ public class CreateHandler extends BaseHandler<CallbackContext> {
 
     private ProgressEvent<ResourceModel, CallbackContext> createResource(ResourceModel model) {
         model.setNamedQueryId(createNamedQuery(model));
-        return ProgressEvent.<ResourceModel, CallbackContext>builder()
-                .resourceModel(model)
-                .status(OperationStatus.SUCCESS)
-                .build();
+        return ProgressEvent.defaultSuccessHandler(model);
     }
 
     private String createNamedQuery(final ResourceModel model) {
@@ -53,8 +48,6 @@ public class CreateHandler extends BaseHandler<CallbackContext> {
             throw new CfnGeneralServiceException("createNamedQuery", e);
         } catch (InvalidRequestException e) {
             throw new CfnInvalidRequestException(createNamedQueryRequest.toString(), e);
-        } catch (Throwable e) {
-            throw new RuntimeException(e);
         }
     }
 }
