+++ namedquery/src/main/java/software/amazon/athena/namedquery/DeleteHandler.java
@@ -1,17 +1,15 @@
-package com.amazonaws.athena.namedquery;
-
-import com.amazonaws.cloudformation.exceptions.CfnGeneralServiceException;
-import com.amazonaws.cloudformation.exceptions.CfnInvalidRequestException;
-import com.amazonaws.cloudformation.proxy.AmazonWebServicesClientProxy;
-import com.amazonaws.cloudformation.proxy.Logger;
-import com.amazonaws.cloudformation.proxy.OperationStatus;
-import com.amazonaws.cloudformation.proxy.ProgressEvent;
-import com.amazonaws.cloudformation.proxy.ResourceHandlerRequest;
+package software.amazon.athena.namedquery;
 
 import software.amazon.awssdk.services.athena.AthenaClient;
 import software.amazon.awssdk.services.athena.model.DeleteNamedQueryRequest;
 import software.amazon.awssdk.services.athena.model.InternalServerException;
 import software.amazon.awssdk.services.athena.model.InvalidRequestException;
+import software.amazon.cloudformation.exceptions.CfnGeneralServiceException;
+import software.amazon.cloudformation.exceptions.CfnInvalidRequestException;
+import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
+import software.amazon.cloudformation.proxy.Logger;
+import software.amazon.cloudformation.proxy.ProgressEvent;
+import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
 
 public class DeleteHandler extends BaseHandler<CallbackContext> {
     private AmazonWebServicesClientProxy clientProxy;
@@ -32,10 +30,7 @@ public class DeleteHandler extends BaseHandler<CallbackContext> {
 
     private ProgressEvent<ResourceModel, CallbackContext> deleteResource(ResourceModel model) {
         deleteNamedQuery(model);
-        return ProgressEvent.<ResourceModel, CallbackContext>builder()
-                .resourceModel(model)
-                .status(OperationStatus.SUCCESS)
-                .build();
+        return ProgressEvent.defaultSuccessHandler(model);
     }
 
     private void deleteNamedQuery(final ResourceModel model) {
@@ -48,8 +43,6 @@ public class DeleteHandler extends BaseHandler<CallbackContext> {
             throw new CfnGeneralServiceException("deleteNamedQuery", e);
         } catch (InvalidRequestException e) {
             throw new CfnInvalidRequestException(deleteNamedQueryRequest.toString(), e);
-        } catch (Throwable e) {
-            throw new RuntimeException(e);
         }
     }
 }
