+++ namedquery/src/main/java/software/amazon/athena/namedquery/ListHandler.java
@@ -1,12 +1,4 @@
-package com.amazonaws.athena.namedquery;
-
-import com.amazonaws.cloudformation.exceptions.CfnGeneralServiceException;
-import com.amazonaws.cloudformation.exceptions.CfnInvalidRequestException;
-import com.amazonaws.cloudformation.proxy.AmazonWebServicesClientProxy;
-import com.amazonaws.cloudformation.proxy.Logger;
-import com.amazonaws.cloudformation.proxy.ProgressEvent;
-import com.amazonaws.cloudformation.proxy.OperationStatus;
-import com.amazonaws.cloudformation.proxy.ResourceHandlerRequest;
+package software.amazon.athena.namedquery;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -16,6 +8,13 @@ import software.amazon.awssdk.services.athena.model.InternalServerException;
 import software.amazon.awssdk.services.athena.model.InvalidRequestException;
 import software.amazon.awssdk.services.athena.model.ListNamedQueriesRequest;
 import software.amazon.awssdk.services.athena.model.ListNamedQueriesResponse;
+import software.amazon.cloudformation.exceptions.CfnGeneralServiceException;
+import software.amazon.cloudformation.exceptions.CfnInvalidRequestException;
+import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
+import software.amazon.cloudformation.proxy.Logger;
+import software.amazon.cloudformation.proxy.OperationStatus;
+import software.amazon.cloudformation.proxy.ProgressEvent;
+import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
 
 public class ListHandler extends BaseHandler<CallbackContext> {
     private AmazonWebServicesClientProxy clientProxy;
@@ -60,8 +59,6 @@ public class ListHandler extends BaseHandler<CallbackContext> {
             throw new CfnGeneralServiceException("listNamedQueriesRequest", e);
         } catch (InvalidRequestException e) {
             throw new CfnInvalidRequestException(listNamedQueriesRequest.toString(), e);
-        } catch (Throwable e) {
-            throw new RuntimeException(e);
         }
     }
 }
