+++ namedquery/src/main/java/software/amazon/athena/namedquery/ReadHandler.java
@@ -1,18 +1,16 @@
-package com.amazonaws.athena.namedquery;
-
-import com.amazonaws.cloudformation.exceptions.CfnGeneralServiceException;
-import com.amazonaws.cloudformation.exceptions.CfnInvalidRequestException;
-import com.amazonaws.cloudformation.proxy.AmazonWebServicesClientProxy;
-import com.amazonaws.cloudformation.proxy.Logger;
-import com.amazonaws.cloudformation.proxy.ProgressEvent;
-import com.amazonaws.cloudformation.proxy.OperationStatus;
-import com.amazonaws.cloudformation.proxy.ResourceHandlerRequest;
+package software.amazon.athena.namedquery;
 
 import software.amazon.awssdk.services.athena.AthenaClient;
 import software.amazon.awssdk.services.athena.model.GetNamedQueryRequest;
 import software.amazon.awssdk.services.athena.model.InternalServerException;
 import software.amazon.awssdk.services.athena.model.InvalidRequestException;
 import software.amazon.awssdk.services.athena.model.NamedQuery;
+import software.amazon.cloudformation.exceptions.CfnGeneralServiceException;
+import software.amazon.cloudformation.exceptions.CfnInvalidRequestException;
+import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
+import software.amazon.cloudformation.proxy.Logger;
+import software.amazon.cloudformation.proxy.ProgressEvent;
+import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
 
 public class ReadHandler extends BaseHandler<CallbackContext> {
     private AmazonWebServicesClientProxy clientProxy;
@@ -30,10 +28,7 @@ public class ReadHandler extends BaseHandler<CallbackContext> {
         clientProxy = proxy;
         athenaClient = AthenaClient.create();
 
-        return ProgressEvent.<ResourceModel, CallbackContext>builder()
-            .resourceModel(getNamedQuery(model))
-            .status(OperationStatus.SUCCESS)
-            .build();
+        return ProgressEvent.defaultSuccessHandler(getNamedQuery(model));
     }
 
     private ResourceModel getNamedQuery(final ResourceModel model) {
@@ -55,8 +50,6 @@ public class ReadHandler extends BaseHandler<CallbackContext> {
             throw new CfnGeneralServiceException("getNamedQuery", e);
         } catch (InvalidRequestException e) {
             throw new CfnInvalidRequestException(getNamedQueryRequest.toString(), e);
-        } catch (Throwable e) {
-            throw new RuntimeException(e);
         }
     }
 }
