+++ namedquery/src/main/java/software/amazon/athena/namedquery/UpdateHandler.java
@@ -1,0 +1,24 @@
+package software.amazon.athena.namedquery;
+
+import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
+import software.amazon.cloudformation.proxy.HandlerErrorCode;
+import software.amazon.cloudformation.proxy.Logger;
+import software.amazon.cloudformation.proxy.OperationStatus;
+import software.amazon.cloudformation.proxy.ProgressEvent;
+import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
+
+public class UpdateHandler extends BaseHandler<CallbackContext> {
+
+    @Override
+    public ProgressEvent<ResourceModel, CallbackContext> handleRequest(
+        final AmazonWebServicesClientProxy proxy,
+        final ResourceHandlerRequest<ResourceModel> request,
+        final CallbackContext callbackContext,
+        final Logger logger) {
+
+        return ProgressEvent.<ResourceModel, CallbackContext>builder()
+            .errorCode(HandlerErrorCode.NotUpdatable)
+            .status(OperationStatus.FAILED)
+            .build();
+    }
+}
