--- namedquery/src/test/java/com/amazonaws/athena/namedquery/UpdateHandlerTest.java
@@ -1,32 +1,0 @@
-package com.amazonaws.athena.namedquery;
-
-import com.amazonaws.cloudformation.exceptions.CfnNotUpdatableException;
-import com.amazonaws.cloudformation.proxy.AmazonWebServicesClientProxy;
-import com.amazonaws.cloudformation.proxy.Logger;
-import com.amazonaws.cloudformation.proxy.ResourceHandlerRequest;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
-@ExtendWith(MockitoExtension.class)
-class UpdateHandlerTest {
-
-    @Mock
-    private AmazonWebServicesClientProxy proxy;
-
-    @Mock
-    private Logger logger;
-
-    @Test
-    void handleRequest_SimpleSuccess() {
-        final ResourceHandlerRequest<ResourceModel> request = ResourceHandlerRequest.<ResourceModel>builder()
-            .desiredResourceState(ResourceModel.builder().build())
-            .build();
-
-        assertThrows(CfnNotUpdatableException.class, () ->
-                new UpdateHandler().handleRequest(proxy, request, null, logger));
-    }
-}
