+++ namedquery/src/test/java/software/amazon/athena/namedquery/CreateHandlerTest.java
@@ -1,12 +1,4 @@
-package com.amazonaws.athena.namedquery;
-
-import com.amazonaws.cloudformation.exceptions.CfnGeneralServiceException;
-import com.amazonaws.cloudformation.exceptions.CfnInvalidRequestException;
-import com.amazonaws.cloudformation.proxy.AmazonWebServicesClientProxy;
-import com.amazonaws.cloudformation.proxy.Logger;
-import com.amazonaws.cloudformation.proxy.OperationStatus;
-import com.amazonaws.cloudformation.proxy.ProgressEvent;
-import com.amazonaws.cloudformation.proxy.ResourceHandlerRequest;
+package software.amazon.athena.namedquery;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -19,13 +11,19 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.doThrow;
 
-import software.amazon.awssdk.services.athena.model.GetNamedQueryResponse;
+import software.amazon.awssdk.services.athena.model.CreateNamedQueryResponse;
 import software.amazon.awssdk.services.athena.model.InternalServerException;
 import software.amazon.awssdk.services.athena.model.InvalidRequestException;
-import software.amazon.awssdk.services.athena.model.NamedQuery;
+import software.amazon.cloudformation.exceptions.CfnGeneralServiceException;
+import software.amazon.cloudformation.exceptions.CfnInvalidRequestException;
+import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
+import software.amazon.cloudformation.proxy.Logger;
+import software.amazon.cloudformation.proxy.OperationStatus;
+import software.amazon.cloudformation.proxy.ProgressEvent;
+import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
 
 @ExtendWith(MockitoExtension.class)
-class ReadHandlerTest {
+class CreateHandlerTest {
     @Mock
     private AmazonWebServicesClientProxy proxy;
 
@@ -36,38 +34,29 @@ class ReadHandlerTest {
     void testSuccessState() {
         // Prepare inputs
         final ResourceModel resourceModel = ResourceModel.builder()
-                .namedQueryId("namedQueryId")
+                .name("name")
+                .database("database")
+                .queryString("queryString")
                 .build();
         final ResourceHandlerRequest<ResourceModel> request = ResourceHandlerRequest.<ResourceModel>builder()
                 .desiredResourceState(resourceModel)
                 .build();
-        final NamedQuery namedQuery = NamedQuery.builder()
-                .database("database")
-                .description("description")
-                .name("name")
-                .queryString("querystring")
-                .build();
+        final String namedQueryId = "namedQueryId";
 
         // Mock
-        doReturn(
-                GetNamedQueryResponse.builder()
-                        .namedQuery(namedQuery)
-                        .build())
+        doReturn(CreateNamedQueryResponse.builder().namedQueryId(namedQueryId).build())
                 .when(proxy)
                 .injectCredentialsAndInvokeV2(any(), any());
 
         // Call
         final ProgressEvent<ResourceModel, CallbackContext> response
-                = new ReadHandler().handleRequest(proxy, request, null, logger);
+                        = new CreateHandler().handleRequest(proxy, request, null, logger);
 
         // Assert
         assertThat(response.getStatus()).isEqualTo(OperationStatus.SUCCESS);
         assertThat(response.getCallbackContext()).isNull();
         assertThat(response.getCallbackDelaySeconds()).isEqualTo(0);
-        assertThat(response.getResourceModel().getDatabase()).isEqualTo(namedQuery.database());
-        assertThat(response.getResourceModel().getDescription()).isEqualTo(namedQuery.description());
-        assertThat(response.getResourceModel().getName()).isEqualTo(namedQuery.name());
-        assertThat(response.getResourceModel().getQueryString()).isEqualTo(namedQuery.queryString());
+        assertThat(response.getResourceModel().getNamedQueryId()).isEqualTo(namedQueryId);
         assertThat(response.getMessage()).isNull();
         assertThat(response.getErrorCode()).isNull();
     }
@@ -76,7 +65,9 @@ class ReadHandlerTest {
     void testInternalServerException() {
         // Prepare inputs
         final ResourceModel resourceModel = ResourceModel.builder()
-                .namedQueryId("namedQueryId")
+                .name("name")
+                .database("database")
+                .queryString("queryString")
                 .build();
         final ResourceHandlerRequest<ResourceModel> request = ResourceHandlerRequest.<ResourceModel>builder()
                 .desiredResourceState(resourceModel)
@@ -89,14 +80,16 @@ class ReadHandlerTest {
 
         // Call
         assertThrows(CfnGeneralServiceException.class, () ->
-                new ReadHandler().handleRequest(proxy, request, null, logger));
+                new CreateHandler().handleRequest(proxy, request, null, logger));
     }
 
     @Test
     void testInvalidRequestException() {
         // Prepare inputs
         final ResourceModel resourceModel = ResourceModel.builder()
-                .namedQueryId("namedQueryId")
+                .name("name")
+                .database("database")
+                .queryString("queryString")
                 .build();
         final ResourceHandlerRequest<ResourceModel> request = ResourceHandlerRequest.<ResourceModel>builder()
                 .desiredResourceState(resourceModel)
@@ -109,26 +102,7 @@ class ReadHandlerTest {
 
         // Call
         assertThrows(CfnInvalidRequestException.class, () ->
-                new ReadHandler().handleRequest(proxy, request, null, logger));
+                new CreateHandler().handleRequest(proxy, request, null, logger));
     }
 
-    @Test
-    void testThrowable() {
-        // Prepare inputs
-        final ResourceModel resourceModel = ResourceModel.builder()
-                .namedQueryId("namedQueryId")
-                .build();
-        final ResourceHandlerRequest<ResourceModel> request = ResourceHandlerRequest.<ResourceModel>builder()
-                .desiredResourceState(resourceModel)
-                .build();
-
-        // Mock
-        doThrow(new RuntimeException())
-                .when(proxy)
-                .injectCredentialsAndInvokeV2(any(), any());
-
-        // Call
-        assertThrows(RuntimeException.class, () ->
-                new ReadHandler().handleRequest(proxy, request, null, logger));
-    }
 }
