+++ namedquery/src/test/java/software/amazon/athena/namedquery/ListHandlerTest.java
@@ -1,0 +1,103 @@
+package software.amazon.athena.namedquery;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.doThrow;
+
+import java.util.Arrays;
+import java.util.List;
+
+import software.amazon.awssdk.services.athena.model.InternalServerException;
+import software.amazon.awssdk.services.athena.model.InvalidRequestException;
+import software.amazon.awssdk.services.athena.model.ListNamedQueriesResponse;
+import software.amazon.cloudformation.exceptions.CfnGeneralServiceException;
+import software.amazon.cloudformation.exceptions.CfnInvalidRequestException;
+import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
+import software.amazon.cloudformation.proxy.Logger;
+import software.amazon.cloudformation.proxy.OperationStatus;
+import software.amazon.cloudformation.proxy.ProgressEvent;
+import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
+
+@ExtendWith(MockitoExtension.class)
+class ListHandlerTest {
+
+    @Mock
+    private AmazonWebServicesClientProxy proxy;
+
+    @Mock
+    private Logger logger;
+
+    @Test
+    void testSuccessState() {
+        // Prepare inputs
+        final ResourceHandlerRequest<ResourceModel> request = ResourceHandlerRequest.<ResourceModel>builder()
+                .desiredResourceState(ResourceModel.builder().build())
+                .build();
+        final List<String> namedQueryIds = Arrays.asList("id1", "id2");
+
+        // Mock
+        doReturn(
+                ListNamedQueriesResponse.builder()
+                        .nextToken("nextToken")
+                        .namedQueryIds(namedQueryIds)
+                        .build())
+                .when(proxy)
+                .injectCredentialsAndInvokeV2(any(), any());
+
+        // Call
+        final ProgressEvent<ResourceModel, CallbackContext> response
+                = new ListHandler().handleRequest(proxy, request, null, logger);
+
+        // Assert
+        assertThat(response.getStatus()).isEqualTo(OperationStatus.SUCCESS);
+        assertThat(response.getCallbackContext()).isNull();
+        assertThat(response.getCallbackDelaySeconds()).isEqualTo(0);
+        assertThat(response.getResourceModels().size()).isEqualTo(namedQueryIds.size());
+        assertThat(response.getResourceModels().get(0).getNamedQueryId()).isEqualTo(namedQueryIds.get(0));
+        assertThat(response.getResourceModels().get(1).getNamedQueryId()).isEqualTo(namedQueryIds.get(1));
+        assertThat(response.getMessage()).isNull();
+        assertThat(response.getErrorCode()).isNull();
+    }
+
+    @Test
+    void testInternalServerException() {
+        // Prepare inputs
+        final ResourceHandlerRequest<ResourceModel> request = ResourceHandlerRequest.<ResourceModel>builder()
+                .desiredResourceState(ResourceModel.builder().build())
+                .build();
+
+        // Mock
+        doThrow(InternalServerException.builder().build())
+                .when(proxy)
+                .injectCredentialsAndInvokeV2(any(), any());
+
+        // Call
+        assertThrows(CfnGeneralServiceException.class, () ->
+                new ListHandler().handleRequest(proxy, request, null, logger));
+    }
+
+    @Test
+    void testInvalidRequestException() {
+        // Prepare inputs
+        final ResourceHandlerRequest<ResourceModel> request = ResourceHandlerRequest.<ResourceModel>builder()
+                .desiredResourceState(ResourceModel.builder().build())
+                .build();
+
+        // Mock
+        doThrow(InvalidRequestException.builder().build())
+                .when(proxy)
+                .injectCredentialsAndInvokeV2(any(), any());
+
+        // Call
+        assertThrows(CfnInvalidRequestException.class, () ->
+                new ListHandler().handleRequest(proxy, request, null, logger));
+    }
+
+}
