+++ namedquery/src/test/java/software/amazon/athena/namedquery/ReadHandlerTest.java
@@ -1,12 +1,4 @@
-package com.amazonaws.athena.namedquery;
-
-import com.amazonaws.cloudformation.exceptions.CfnGeneralServiceException;
-import com.amazonaws.cloudformation.exceptions.CfnInvalidRequestException;
-import com.amazonaws.cloudformation.proxy.AmazonWebServicesClientProxy;
-import com.amazonaws.cloudformation.proxy.Logger;
-import com.amazonaws.cloudformation.proxy.OperationStatus;
-import com.amazonaws.cloudformation.proxy.ProgressEvent;
-import com.amazonaws.cloudformation.proxy.ResourceHandlerRequest;
+package software.amazon.athena.namedquery;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -19,12 +11,20 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.doThrow;
 
-import software.amazon.awssdk.services.athena.model.DeleteNamedQueryResponse;
+import software.amazon.awssdk.services.athena.model.GetNamedQueryResponse;
 import software.amazon.awssdk.services.athena.model.InternalServerException;
 import software.amazon.awssdk.services.athena.model.InvalidRequestException;
+import software.amazon.awssdk.services.athena.model.NamedQuery;
+import software.amazon.cloudformation.exceptions.CfnGeneralServiceException;
+import software.amazon.cloudformation.exceptions.CfnInvalidRequestException;
+import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
+import software.amazon.cloudformation.proxy.Logger;
+import software.amazon.cloudformation.proxy.OperationStatus;
+import software.amazon.cloudformation.proxy.ProgressEvent;
+import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
 
 @ExtendWith(MockitoExtension.class)
-class DeleteHandlerTest {
+class ReadHandlerTest {
     @Mock
     private AmazonWebServicesClientProxy proxy;
 
@@ -40,22 +40,33 @@ class DeleteHandlerTest {
         final ResourceHandlerRequest<ResourceModel> request = ResourceHandlerRequest.<ResourceModel>builder()
                 .desiredResourceState(resourceModel)
                 .build();
+        final NamedQuery namedQuery = NamedQuery.builder()
+                .database("database")
+                .description("description")
+                .name("name")
+                .queryString("querystring")
+                .build();
 
         // Mock
-        doReturn(DeleteNamedQueryResponse.builder().build())
+        doReturn(
+                GetNamedQueryResponse.builder()
+                        .namedQuery(namedQuery)
+                        .build())
                 .when(proxy)
                 .injectCredentialsAndInvokeV2(any(), any());
 
         // Call
         final ProgressEvent<ResourceModel, CallbackContext> response
-                = new DeleteHandler().handleRequest(proxy, request, null, logger);
+                = new ReadHandler().handleRequest(proxy, request, null, logger);
 
         // Assert
         assertThat(response.getStatus()).isEqualTo(OperationStatus.SUCCESS);
         assertThat(response.getCallbackContext()).isNull();
         assertThat(response.getCallbackDelaySeconds()).isEqualTo(0);
-        assertThat(response.getResourceModel()).isEqualTo(request.getDesiredResourceState());
-        assertThat(response.getResourceModels()).isNull();
+        assertThat(response.getResourceModel().getDatabase()).isEqualTo(namedQuery.database());
+        assertThat(response.getResourceModel().getDescription()).isEqualTo(namedQuery.description());
+        assertThat(response.getResourceModel().getName()).isEqualTo(namedQuery.name());
+        assertThat(response.getResourceModel().getQueryString()).isEqualTo(namedQuery.queryString());
         assertThat(response.getMessage()).isNull();
         assertThat(response.getErrorCode()).isNull();
     }
@@ -77,7 +88,7 @@ class DeleteHandlerTest {
 
         // Call
         assertThrows(CfnGeneralServiceException.class, () ->
-                new DeleteHandler().handleRequest(proxy, request, null, logger));
+                new ReadHandler().handleRequest(proxy, request, null, logger));
     }
 
     @Test
@@ -97,27 +108,6 @@ class DeleteHandlerTest {
 
         // Call
         assertThrows(CfnInvalidRequestException.class, () ->
-                new DeleteHandler().handleRequest(proxy, request, null, logger));
+                new ReadHandler().handleRequest(proxy, request, null, logger));
     }
-
-    @Test
-    void testThrowable() {
-        // Prepare inputs
-        final ResourceModel resourceModel = ResourceModel.builder()
-                .namedQueryId("namedQueryId")
-                .build();
-        final ResourceHandlerRequest<ResourceModel> request = ResourceHandlerRequest.<ResourceModel>builder()
-                .desiredResourceState(resourceModel)
-                .build();
-
-        // Mock
-        doThrow(new RuntimeException())
-                .when(proxy)
-                .injectCredentialsAndInvokeV2(any(), any());
-
-        // Call
-        assertThrows(RuntimeException.class, () ->
-                new DeleteHandler().handleRequest(proxy, request, null, logger));
-    }
-
 }
